chore(mpt): noop provider refactor
