+++ crates/mpt/src/fetcher.rs
@@ -1,7 +1,7 @@
 //! Contains the [TrieProvider] trait for fetching trie node preimages, contract bytecode, and
 //! headers.
 
-use alloc::string::{String, ToString};
+use alloc::string::ToString;
 use alloy_consensus::Header;
 use alloy_primitives::{Address, Bytes, B256, U256};
 use core::fmt::Display;
@@ -92,49 +92,4 @@ pub trait TrieHinter {
         slot: U256,
         block_number: u64,
     ) -> Result<(), Self::Error>;
-}
-
-/// The default, no-op implementation of the [TrieProvider] trait, used for testing.
-#[derive(Debug, Clone, Copy)]
-pub struct NoopTrieProvider;
-
-impl TrieProvider for NoopTrieProvider {
-    type Error = String;
-
-    fn trie_node_preimage(&self, _key: B256) -> Result<Bytes, Self::Error> {
-        Ok(Bytes::new())
-    }
-
-    fn bytecode_by_hash(&self, _code_hash: B256) -> Result<Bytes, Self::Error> {
-        Ok(Bytes::new())
-    }
-
-    fn header_by_hash(&self, _hash: B256) -> Result<Header, Self::Error> {
-        Ok(Header::default())
-    }
-}
-
-/// The default, no-op implementation of the [TrieHinter] trait, used for testing.
-#[derive(Debug, Clone, Copy)]
-pub struct NoopTrieHinter;
-
-impl TrieHinter for NoopTrieHinter {
-    type Error = String;
-
-    fn hint_trie_node(&self, _hash: B256) -> Result<(), Self::Error> {
-        Ok(())
-    }
-
-    fn hint_account_proof(&self, _address: Address, _block_number: u64) -> Result<(), Self::Error> {
-        Ok(())
-    }
-
-    fn hint_storage_proof(
-        &self,
-        _address: Address,
-        _slot: U256,
-        _block_number: u64,
-    ) -> Result<(), Self::Error> {
-        Ok(())
-    }
 }
